requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
analytics-python==1.4.0
anyio==3.6.1
appnope==0.1.2
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backoff==1.10.0
bcrypt==3.2.2
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
distlib==0.3.4
#dlib==19.18.0
executing==0.8.2
face-alignment==1.3.5
fastapi==0.78.0
ffmpy==0.3.0
filelock==3.4.2
fire==0.4.0
flake8==4.0.1
fonttools==4.33.3
frozenlist==1.3.0
gradio==3.0.10
h11==0.13.0
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
langid==1.1.6
linkify-it-py==1.0.3
llvmlite==0.38.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mdit-py-plugins==0.3.0
mdurl==0.1.1
monotonic==1.6
multidict==6.0.2
mypy-extensions==0.4.3
numba==0.55.2
numpy
opencv-python==4.5.5.64
orjson==3.6.8
overrides==4.1.2
packaging==21.3
pandas
paramiko==2.11.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.4.1
prompt-toolkit==3.0.24
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.1
pydub==0.25.1
pyflakes==2.4.0
Pygments==2.11.2
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
PyWavelets==1.3.0
#PyYAML==3.13
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scikit-video
scipy==1.7.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.1.3
starlette==0.19.1
tabulate==0.8.9
termcolor==1.1.0
threadpoolctl==3.0.0
tomli==1.2.3
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.3
traitlets==5.1.1
typing-utils==0.1.0
typing_extensions==4.0.1
uc-micro-py==1.0.1
Unidecode==1.3.2
urllib3==1.26.8
uvicorn==0.17.6
virtualenv==20.13.0
wcwidth==0.2.5
yarl==1.7.2
google-cloud-storage
fastapi
uvicorn
python-multipart
#cmake
